+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   handling for `cast/2` and `send_frame/2`.
 - Add access to response headers during `handle_connect/2` via `Conn.resp_headers`.
 - Add `Conn.parse_url/1` to handle url to URI conversion.
+- Allow `Conn.new/2` to use a url string instead of a `URI` struct.
 - Automatically add a "/" path to a pathless url.
   - The HTTP request will break without a valid path!